backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.1
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
pydantic==2.5.0
